fix: keep SSR session loading in the Nitro auth instance
